+++ mayan/apps/common/management/commands/initialsetup.py
@@ -95,7 +95,7 @@ class Command(management.BaseCommand):
         if not options.get('no_dependencies', False):
             management.call_command(command_name='installdependencies')
             management.call_command(
-                command_name='preparestatic', interactive=False
+                command_name='appearance_prepare_static', interactive=False
             )
 
         management.call_command(command_name='createautoadmin')