+++ mayan/apps/common/management/commands/performupgrade.py
@@ -51,7 +51,7 @@ class Command(management.BaseCommand):
                 command_name='installdependencies'
             )
             management.call_command(
-                command_name='preparestatic', interactive=False
+                command_name='appearance_prepare_static', interactive=False
             )
 
         try: