+++ mayan/apps/task_manager/tests/test_management_commands.py
@@ -10,13 +10,15 @@ from .mixins import TaskManagerManagementCommandTestMixin
 class TaskManagerManagementCommandTestCase(
     TaskManagerManagementCommandTestMixin, BaseTestCase
 ):
-    def test_purgeperiodictasks_management_command(self):
+    def test_purge_periodic_tasks_management_command(self):
         self._test_interval_schedule_count = IntervalSchedule.objects.count()
         self._test_periodic_task_count = PeriodicTask.objects.count()
 
         self._clear_events()
 
-        management.call_command(command_name='purgeperiodictasks')
+        management.call_command(
+            command_name='task_manager_purge_periodic_tasks'
+        )
 
         self.assertEqual(
             IntervalSchedule.objects.count(),