+++ mayan/apps/autoadmin/dependencies.py
@@ -1,5 +1,5 @@
 from mayan.apps.dependencies.classes import PythonDependency
 
 PythonDependency(
-    module=__name__, name='django-solo', version_string='==1.1.5'
+    module=__name__, name='django-solo', version_string='==2.0.0'
 )