+++ mayan/apps/cabinets/dependencies.py
@@ -1,5 +1,5 @@
 from mayan.apps.dependencies.classes import JavaScriptDependency
 
 JavaScriptDependency(
-    module=__name__, name='jstree', version_string='=3.3.11'
+    module=__name__, name='jstree', version_string='=3.3.12'
 )