+++ mayan/apps/dependencies/dependencies.py
@@ -1,8 +1,8 @@
 from .classes import PythonDependency
 
 PythonDependency(
-    module=__name__, name='node-semver', version_string='==0.8.0'
+    module=__name__, name='node-semver', version_string='==0.8.1'
 )
 PythonDependency(
-    module=__name__, name='packaging', version_string='==21.0'
+    module=__name__, name='packaging', version_string='==21.3'
 )