+++ mayan/apps/django_gpg/dependencies.py
@@ -37,5 +37,5 @@ PythonDependency(
         LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
         OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
         ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-    ''', module=__name__, name='python_gnupg', version_string='==0.4.7'
+    ''', module=__name__, name='python_gnupg', version_string='==0.4.8'
 )