+++ mayan/apps/dynamic_search/dependencies.py
@@ -4,7 +4,7 @@ PythonDependency(
     module=__name__, name='Whoosh', version_string='==2.7.4'
 )
 PythonDependency(
-    module=__name__, name='elasticsearch', version_string='==7.16.0'
+    module=__name__, name='elasticsearch', version_string='==7.17.1'
 )
 PythonDependency(
     module=__name__, name='elasticsearch-dsl', version_string='==7.4.0'