+++ mayan/apps/mime_types/dependencies.py
@@ -40,5 +40,5 @@ PythonDependency(
         LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
         OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
         SOFTWARE.
-    ''', module=__name__, name='python-magic', version_string='==0.4.24'
+    ''', module=__name__, name='python-magic', version_string='==0.4.25'
 )