+++ mayan/apps/platform/dependencies.py
@@ -1,14 +1,14 @@
 from mayan.apps.dependencies.classes import PythonDependency
 
 PythonDependency(
-    module=__name__, name='gevent', version_string='==21.8.0'
+    module=__name__, name='gevent', version_string='==21.12.0'
 )
 PythonDependency(
     module=__name__, name='gunicorn', version_string='==20.1.0'
 )
 PythonDependency(
-    module=__name__, name='sentry-sdk', version_string='==1.4.1'
+    module=__name__, name='sentry-sdk', version_string='==1.5.8'
 )
 PythonDependency(
-    module=__name__, name='whitenoise', version_string='==5.3.0'
+    module=__name__, name='whitenoise', version_string='==6.0.0'
 )