+++ mayan/apps/rest_api/dependencies.py
@@ -23,7 +23,8 @@ PythonDependency(
         LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
         OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
         WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-    ''', module=__name__, name='django-cors-headers', version_string='==3.8.0'
+    ''', module=__name__, name='django-cors-headers',
+    version_string='==3.10.0'
 )
 PythonDependency(
     copyright_text='''
@@ -34,7 +35,8 @@ PythonDependency(
         Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
 
         THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-    ''', module=__name__, name='djangorestframework', version_string='==3.12.4'
+    ''', module=__name__, name='djangorestframework',
+    version_string='==3.13.1'
 )
 PythonDependency(
     copyright_text='''
@@ -61,8 +63,8 @@ PythonDependency(
     module=__name__, name='flex', version_string='==6.14.1'
 )
 PythonDependency(
-    module=__name__, name='jsonschema', version_string='==3.2.0'
+    module=__name__, name='jsonschema', version_string='==4.4.0'
 )
 PythonDependency(
-    module=__name__, name='swagger-spec-validator', version_string='==2.7.3'
+    module=__name__, name='swagger-spec-validator', version_string='==2.7.4'
 )